refactor: organize text processing tests
